courses/lis500/ml-project/requirements.txt
tensorflow
pillow
numpy
scikit-learn
matplotlib
tensorflowjs
opencv-python
pillow-heif
